+++ rds/src/v3/RdsClient.cpp
@@ -3516,6 +3516,39 @@ std::shared_ptr<ListUpgradeHistoriesResponse> RdsClient::listUpgradeHistories(Li
 
     return localVarResult;
 }
+std::shared_ptr<ListVolumeInfoResponse> RdsClient::listVolumeInfo(ListVolumeInfoRequest &request)
+{
+    std::string localVarPath = "/v3/{project_id}/instances/{instance_id}/volumes";
+
+    std::map<std::string, std::string> localVarQueryParams;
+    std::map<std::string, std::string> localVarHeaderParams;
+    std::map<std::string, std::string> localVarFormParams;
+    std::map<std::string, std::string> localVarPathParams;
+
+    localVarPathParams["instance_id"] = parameterToString(request.getInstanceId());
+
+    bool isJson = false;
+    bool isMultiPart = false;
+    bool isBson = false;
+    std::string contentType = getContentType("application/json", isJson, isMultiPart, isBson);
+    localVarHeaderParams["Content-Type"] = contentType;
+
+    if (request.xLanguageIsSet()) {
+        localVarHeaderParams["X-Language"] = parameterToString(request.getXLanguage());
+    }
+
+    std::string localVarHttpBody;
+
+    std::unique_ptr<HttpResponse> res = callApi("GET", localVarPath, localVarPathParams, localVarQueryParams,
+        localVarHeaderParams, localVarHttpBody, RdsMeta::genRequestDefForListVolumeInfo());
+
+    std::shared_ptr<ListVolumeInfoResponse> localVarResult = std::make_shared<ListVolumeInfoResponse>();
+    localVarResult->setStatusCode(res->getStatusCode());
+    localVarResult->setHeaderParams(res->getHeaderParams());
+    localVarResult->setHttpBody(res->getHttpBody());
+
+    return localVarResult;
+}
 std::shared_ptr<ListXellogFilesResponse> RdsClient::listXellogFiles(ListXellogFilesRequest &request)
 {
     std::string localVarPath = "/v3/{project_id}/instances/{instance_id}/xellog-files";